Moved rodata from C to GLOBAL_ASM in lib/unknown_0D3160
